# engine: speedup loadXML methods by using QStringLiteral