--- a/engine/src/channelsgroup.h
+++ b/engine/src/channelsgroup.h
@@ -32,7 +32,7 @@ class Doc;
  * @{
  */
 
-#define KXMLQLCChannelsGroup QString("ChannelsGroup")
+#define KXMLQLCChannelsGroup QStringLiteral("ChannelsGroup")
 
 class ChannelsGroup : public QObject
 {