--- a/engine/src/cuestack.h
+++ b/engine/src/cuestack.h
@@ -41,12 +41,12 @@ class Doc;
  * @{
  */
 
-#define KXMLQLCCueStack              QString("CueStack")
-#define KXMLQLCCueStackID            QString("ID")
-#define KXMLQLCCueStackSpeed         QString("Speed")
-#define KXMLQLCCueStackSpeedFadeIn   QString("FadeIn")
-#define KXMLQLCCueStackSpeedFadeOut  QString("FadeOut")
-#define KXMLQLCCueStackSpeedDuration QString("Duration")
+#define KXMLQLCCueStack              QStringLiteral("CueStack")
+#define KXMLQLCCueStackID            QStringLiteral("ID")
+#define KXMLQLCCueStackSpeed         QStringLiteral("Speed")
+#define KXMLQLCCueStackSpeedFadeIn   QStringLiteral("FadeIn")
+#define KXMLQLCCueStackSpeedFadeOut  QStringLiteral("FadeOut")
+#define KXMLQLCCueStackSpeedDuration QStringLiteral("Duration")
 
 class CueStack : public QObject, public DMXSource
 {