… engine/src/efx.h
… engine/src/qlcchannel.cpp
… engine/src/qlcinputchannel.h
… qmlui/virtualconsole/vcanimation.h
… qmlui/virtualconsole/vcaudiotrigger.h
… qmlui/virtualconsole/vcbutton.h
… qmlui/virtualconsole/vcclock.h
… qmlui/virtualconsole/vccuelist.h
… qmlui/virtualconsole/vcframe.h
… qmlui/virtualconsole/vclabel.h
… qmlui/virtualconsole/vcslider.h
… qmlui/virtualconsole/vcsoloframe.h
… qmlui/virtualconsole/vcspeeddial.h
… qmlui/virtualconsole/vcwidget.h
… qmlui/virtualconsole/vcxypad.h
… qmlui/virtualconsole/virtualconsole.h
… ui/src/app.h
… ui/src/audiobar.h
… ui/src/fixturemanager.cpp
… ui/src/simpledesk.h
… ui/src/simpledeskengine.h
… ui/src/virtualconsole/vcaudiotriggers.h
… ui/src/virtualconsole/vcbutton.h
… ui/src/virtualconsole/vcclock.h
… ui/src/virtualconsole/vccuelist.h
… ui/src/virtualconsole/vcframe.h
… ui/src/virtualconsole/vcframepageshortcut.h
… ui/src/virtualconsole/vclabel.h
… ui/src/virtualconsole/vcmatrix.h
… ui/src/virtualconsole/vcmatrixcontrol.h