reduce number of temporary allocations when parsing xml

On my old i3-3110M, with 1615 fixtures in ~/.qlcplus/fixtures main/qlcplus -w -p --open ~/test.qxw -platform vnc:mode=websock

Before:
heaptrack reports 1278772 temporary allocations via QString -> QArrayData::allocate2 QLCInputChannel::loadXML took 1.1 seconds accumulative (~30000 calls) QLCInputProfile::loadXML took 35 ms accumulative (38 calls)

After:
heaptrack reports 325342 temporary allocations via QString -> QArrayData::allocate2 QLCInputChannel::loadXML took 0.79 seconds accumulative (~30000 calls) QLCInputProfile::loadXML took 31 ms accumulative (38 calls)

On Raspberry Pi 4, with only 3 custom fixtures
main/qlcplus -w -p --open ~/test.qxw -platform vnc:mode=websock

Before:
QLCInputProfile::loadXML took 94 ms cold start and 57 ms hot start accumulative (38 calls)

After:
QLCInputProfile::loadXML took 89 ms cold start and 53 ms hot start accumulative (38 calls)