@mcallegari 2025-04-14
Hey thanks for this. Shouldn't this be applied to all define + QString cases?
`grep -r define * | grep QString`