@yestalgia 2025-04-14
Keen to see the new numbers 🤩